Penalized precision matrix estimation via ADMM

Penalized precision matrix estimation using the ADMM algorithm. Consider the case where \(X_{1}, ..., X_{n}\) are iid \(N_{p}(\mu, \Sigma)\) and we are tasked with estimating the precision matrix, denoted \(\Omega \equiv \Sigma^{-1}\). This function solves the following optimization problem:

Objective:

\(\hat{\Omega}_{\lambda} = \arg\min_{\Omega \in S_{+}^{p}} \left\{ Tr\left(S\Omega\right) - \log \det\left(\Omega \right) + \lambda\left[\frac{1 - \alpha}{2}\left\| \Omega \right|_{F}^{2} + \alpha\left\| \Omega \right\|_{1} \right] \right\}\)

where \(0 \leq \alpha \leq 1\), \(\lambda > 0\), \(\left\|\cdot \right\|_{F}^{2}\) is the Frobenius norm and we define \(\left\|A \right\|_{1} = \sum_{i, j} \left| A_{ij} \right|\). This elastic net penalty is identical to the penalty used in the popular penalized regression package glmnet. Clearly, when \(\alpha = 0\) the elastic-net reduces to a ridge-type penalty and when \(\alpha = 1\) it reduces to a lasso-type penalty.

ADMMsigma(X = NULL, S = NULL, nlam = 10, lam.min.ratio = 0.01,
  lam = NULL, alpha = seq(0, 1, 0.2), diagonal = FALSE, path = FALSE,
  rho = 2, mu = 10, tau.inc = 2, tau.dec = 2, crit = c("ADMM",
  "loglik"), tol.abs = 1e-04, tol.rel = 1e-04, maxit = 10000,
  adjmaxit = NULL, K = 5, crit.cv = c("loglik", "AIC", "BIC"),
  start = c("warm", "cold"), cores = 1, trace = c("progress", "print",
  "none"))

Arguments

X
option to provide a nxp data matrix. Each row corresponds to a single observation and each column contains n observations of a single feature/variable.

S
option to provide a pxp sample covariance matrix (denominator n). If argument is NULL and X is provided instead then S will be computed automatically.

nlam
number of lam tuning parameters for penalty term generated from lam.min.ratio and lam.max (automatically generated). Defaults to 10.

lam.min.ratio
smallest lam value provided as a fraction of lam.max. The function will automatically generate nlam tuning parameters from lam.min.ratio*lam.max to lam.max in log10 scale. lam.max is calculated to be the smallest lam such that all off-diagonal entries in Omega are equal to zero (alpha = 1). Defaults to 1e-2.

lam
option to provide positive tuning parameters for penalty term. This will cause nlam and lam.min.ratio to be disregarded. If a vector of parameters is provided, they should be in increasing order. Defaults to NULL.

alpha
elastic net mixing parameter contained in [0, 1]. 0 = ridge, 1 = lasso. If a vector of parameters is provided, they should be in increasing order. Defaults to grid of values seq(0, 1, 0.2).

diagonal
option to penalize the diagonal elements of the estimated precision matrix (\(\Omega\)). Defaults to FALSE.

path
option to return the regularization path. This option should be used with extreme care if the dimension is large. If set to TRUE, cores must be set to 1 and errors and optimal tuning parameters will based on the full sample. Defaults to FALSE.

rho
initial step size for ADMM algorithm.

mu
factor for primal and residual norms in the ADMM algorithm. This will be used to adjust the step size rho after each iteration.

tau.inc
factor in which to increase step size rho

tau.dec
factor in which to decrease step size rho

crit
criterion for convergence (ADMM or loglik). If crit = loglik then iterations will stop when the relative change in log-likelihood is less than tol.abs. Default is ADMM and follows the procedure outlined in Boyd, et al.

tol.abs
absolute convergence tolerance. Defaults to 1e-4.

tol.rel
relative convergence tolerance. Defaults to 1e-4.

maxit
maximum number of iterations. Defaults to 1e4.

adjmaxit
adjusted maximum number of iterations. During cross validation this option allows the user to adjust the maximum number of iterations after the first lam tuning parameter has converged (for each alpha). This option is intended to be paired with warm starts and allows for 'one-step' estimators. Defaults to NULL.

K
specify the number of folds for cross validation.

crit.cv
cross validation criterion (loglik, AIC, or BIC). Defaults to loglik.

start
specify warm or cold start for cross validation. Default is warm.

cores
option to run CV in parallel. Defaults to cores = 1.

trace
option to display progress of CV. Choose one of progress to print a progress bar, print to print completed tuning parameters, or none.

Value

returns class object ADMMsigma which includes:

Call

function call.

Iterations

number of iterations.

Tuning

optimal tuning parameters (lam and alpha).

Lambdas

grid of lambda values for CV.

Alphas

grid of alpha values for CV.

maxit

maximum number of iterations.

Omega

estimated penalized precision matrix.

Sigma

estimated covariance matrix from the penalized precision matrix (inverse of Omega).

Path

array containing the solution path. Solutions will be ordered in ascending alpha values for each lambda.

Z

final sparse update of estimated penalized precision matrix.

Y

final dual update.

rho

final step size.

Loglik

penalized log-likelihood for Omega

MIN.error

minimum average cross validation error (cv.crit) for optimal parameters.

AVG.error

average cross validation error (cv.crit) across all folds.

CV.error

cross validation errors (cv.crit).

Details

For details on the implementation of 'ADMMsigma', see the vignette https://mgallow.github.io/ADMMsigma/.

References

Boyd, Stephen, Neal Parikh, Eric Chu, Borja Peleato, Jonathan Eckstein, and others. 2011. 'Distributed Optimization and Statistical Learning via the Alternating Direction Method of Multipliers.' Foundations and Trends in Machine Learning 3 (1). Now Publishers, Inc.: 1-122. https://web.stanford.edu/~boyd/papers/pdf/admm_distr_stats.pdf

Hu, Yue, Chi, Eric C, amd Allen, Genevera I. 2016. 'ADMM Algorithmic Regularization Paths for Sparse Statistical Machine Learning.' Splitting Methods in Communication, Imaging, Science, and Engineering. Springer: 433-459.

Zou, Hui and Hastie, Trevor. 2005. 'Regularization and Variable Selection via the Elastic Net.' Journal of the Royal Statistial Society: Series B (Statistical Methodology) 67 (2). Wiley Online Library: 301-320.

Rothman, Adam. 2017. 'STAT 8931 notes on an algorithm to compute the Lasso-penalized Gaussian likelihood precision matrix estimator.'

See also

Examples

# generate data from a sparse matrix
# first compute covariance matrix
S = matrix(0.7, nrow = 5, ncol = 5)
for (i in 1:5){
 for (j in 1:5){
   S[i, j] = S[i, j]^abs(i - j)
 }
 }

# generate 100 x 5 matrix with rows drawn from iid N_p(0, S)
set.seed(123)
Z = matrix(rnorm(100*5), nrow = 100, ncol = 5)
out = eigen(S, symmetric = TRUE)
S.sqrt = out$vectors %*% diag(out$values^0.5)
S.sqrt = S.sqrt %*% t(out$vectors)
X = Z %*% S.sqrt

# elastic-net type penalty (use CV for optimal lambda and alpha)
ADMMsigma(X)

#> 
#> Call: ADMMsigma(X = X)
#> 
#> Iterations: 66
#> 
#> Tuning parameters:
#>       log10(lam)  alpha
#> [1,]      -1.599      1
#> 
#> Log-likelihood: -108.41022
#> 
#> Omega:
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  2.15228 -1.26841  0.00000  0.00000  0.19744
#> [2,] -1.26841  2.78830 -1.31943 -0.08246  0.01018
#> [3,]  0.00000 -1.31943  2.84979 -1.16708 -0.01015
#> [4,]  0.00000 -0.08246 -1.16708  2.49277 -1.18844
#> [5,]  0.19744  0.01018 -0.01015 -1.18844  1.88069

# ridge penalty (use CV for optimal lambda)
ADMMsigma(X, alpha = 0)

#> 
#> Call: ADMMsigma(X = X, alpha = 0)
#> 
#> Iterations: 58
#> 
#> Tuning parameters:
#>       log10(lam)  alpha
#> [1,]      -1.599      0
#> 
#> Log-likelihood: -102.22452
#> 
#> Omega:
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  2.12904 -1.21826 -0.01544 -0.03166  0.23187
#> [2,] -1.21826  2.74648 -1.25141 -0.24670  0.16448
#> [3,] -0.01544 -1.25141  2.76938 -1.00483 -0.19582
#> [4,] -0.03166 -0.24670 -1.00483  2.44699 -1.15723
#> [5,]  0.23187  0.16448 -0.19582 -1.15723  1.90428

# lasso penalty (lam = 0.1)
ADMMsigma(X, lam = 0.1, alpha = 1)

#> 
#> Call: ADMMsigma(X = X, lam = 0.1, alpha = 1)
#> 
#> Iterations: 67
#> 
#> Tuning parameters:
#>       log10(lam)  alpha
#> [1,]          -1      1
#> 
#> Log-likelihood: -140.60864
#> 
#> Omega:
#>          [,1]     [,2]     [,3]     [,4]     [,5]
#> [1,]  1.81633 -0.87377  0.00000  0.00000  0.00000
#> [2,] -0.87377  2.13797 -0.93396 -0.04670  0.00000
#> [3,]  0.00000 -0.93396  2.22066 -0.85016  0.00000
#> [4,]  0.00000 -0.04670 -0.85016  1.95244 -0.85819
#> [5,]  0.00000  0.00000  0.00000 -0.85819  1.59722